backend/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cloudinary==1.44.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
dj-rest-auth==5.0.2
django-allauth==0.57.0
google-auth==2.27.0
google-generativeai==0.8.3
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycryptodome==3.20.0
python-dotenv==1.2.1
PyPDF2==3.0.1
redis==5.0.1
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.11.0